refactor: move auth handlers to package
