+++ handlers/auth/db_alias.go
@@ -1,0 +1,10 @@
+package auth
+
+import db "github.com/arran4/goa4web/internal/db"
+
+type (
+	DBTX                     = db.DBTX                     // used
+	Queries                  = db.Queries                  // used
+	InsertLoginAttemptParams = db.InsertLoginAttemptParams // used
+	User                     = db.User                     // used
+)
