+++ routes.go
@@ -230,14 +230,14 @@ func registerInformationRoutes(r *mux.Router) {
 
 func registerRegisterRoutes(r *mux.Router) {
 	rr := r.PathPrefix("/register").Subrouter()
-	rr.HandleFunc("", registerPage).Methods("GET").MatcherFunc(Not(auth.RequiresAnAccount()))
-	rr.HandleFunc("", registerActionPage).Methods("POST").MatcherFunc(Not(auth.RequiresAnAccount())).MatcherFunc(common.TaskMatcher(TaskRegister))
+	rr.HandleFunc("", auth.RegisterPage).Methods("GET").MatcherFunc(Not(auth.RequiresAnAccount()))
+	rr.HandleFunc("", auth.RegisterActionPage).Methods("POST").MatcherFunc(Not(auth.RequiresAnAccount())).MatcherFunc(common.TaskMatcher(TaskRegister))
 }
 
 func registerLoginRoutes(r *mux.Router) {
 	ulr := r.PathPrefix("/login").Subrouter()
-	ulr.HandleFunc("", loginUserPassPage).Methods("GET").MatcherFunc(Not(auth.RequiresAnAccount()))
-	ulr.HandleFunc("", loginActionPage).Methods("POST").MatcherFunc(Not(auth.RequiresAnAccount())).MatcherFunc(common.TaskMatcher(TaskLogin))
+	ulr.HandleFunc("", auth.LoginUserPassPage).Methods("GET").MatcherFunc(Not(auth.RequiresAnAccount()))
+	ulr.HandleFunc("", auth.LoginActionPage).Methods("POST").MatcherFunc(Not(auth.RequiresAnAccount())).MatcherFunc(common.TaskMatcher(TaskLogin))
 }
 
 func registerAdminRoutes(r *mux.Router) {
